dnsdist: Move libssl.hh to pdns/, it's needed for the fuzz target
